# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards Management API
  version: 1.0.0
  description: |
    REST API для управления банковскими картами, пользователями и переводами между своими картами.

    Основные возможности:
      - JWT-аутентификация и role-based access (`ADMIN`, `USER`)
      - управление картами
      - переводы между своими картами
      - управление пользователями для администратора
      - пагинация, фильтрация, валидация и единый формат ошибок
servers:
  - url: http://localhost:8080/api
    description: Local environment
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Получить список пользователей
      description: Доступно только пользователю с ролью `ADMIN`.
      operationId: getAllUsers
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Поиск по username, email, firstName и lastName
        - name: enabled
          in: query
          required: false
          schema:
            type: boolean
        - name: deleted
          in: query
          required: false
          schema:
            type: boolean
        - name: role
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/RoleName'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Страница пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/{userId}:
    get:
      tags: [Users]
      summary: Получить пользователя по идентификатору
      description: Доступно только пользователю с ролью `ADMIN`.
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/UserIdPathParam'
      responses:
        '200':
          description: Данные пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      tags: [Users]
      summary: Обновить пользователя
      description: |
        Доступно только пользователю с ролью `ADMIN`.
        Собственным аккаунтом через административный эндпоинт управлять нельзя.
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/UserIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Пользователь обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
    delete:
      tags: [Users]
      summary: Удалить пользователя
      description: |
        Доступно только пользователю с ролью `ADMIN`.
        Пользователь помечается как удалённый и отключается.
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/UserIdPathParam'
      responses:
        '204':
          description: Пользователь помечен как удалённый
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserIdPathParam:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Идентификатор пользователя
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Номер страницы, начиная с 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Размер страницы
  responses:
    ValidationError:
      description: Ошибка валидации запроса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    UnauthorizedError:
      description: Необходима аутентификация или передан неверный токен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ForbiddenError:
      description: Недостаточно прав для выполнения операции
    NotFoundError:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ConflictError:
      description: Бизнес-конфликт
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    RoleName:
      type: string
      enum: [ADMIN, USER]
...
    UpdateUserRequest:
      type: object
      properties:
        enabled:
          type: boolean
        deleted:
          type: boolean
        roles:
          type: array
          minItems: 1
          maxItems: 2
          uniqueItems: true
          items:
            $ref: '#/components/schemas/RoleName'
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        enabled:
          type: boolean
        deleted:
          type: boolean
        roles:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, username, email, firstName, lastName, enabled, deleted, roles, createdAt, updatedAt]
    ApiErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        validationErrors:
          type: object
          nullable: true
          additionalProperties:
            type: string
      required: [timestamp, status, error, message, path]
...
    UserPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
      required: [content, page, size, totalElements, totalPages, first, last]